+++ patches/src/main/kotlin/app/morphe/patches/nomone/misc/telemetry/DisableTelemetryPatch.kt
@@ -1,0 +1,16 @@
+package app.morphe.patches.nomone.misc.telemetry
+
+import app.morphe.patcher.patch.bytecodePatch
+import app.morphe.util.returnEarly
+
+@Suppress("unused")
+val disableTelemetryPatch = bytecodePatch(
+    name = "Disable telemetry",
+    description = "Disables event logging sent to the app's custom endpoint."
+) {
+    compatibleWith("nom.vrd"("1.9.3-GooglePlay"))
+
+    execute {
+        SendTelemetryEventFingerprint.method.returnEarly()
+    }
+}