+++ patches/src/main/kotlin/app/morphe/patches/wallcraft/premium/Fingerprints.kt
@@ -1,0 +1,8 @@
+package app.morphe.patches.wallcraft.premium
+
+import app.morphe.patcher.Fingerprint
+
+object GetSubscriptionStateFingerprint : Fingerprint (
+    name = "getSubscriptionStateByOptions",
+    definingClass = "Lcom/wallpaperscraft/billing/core/SubscriptionManager;"
+)